Refactor: Editor 컴포넌트 캡슐화
