+++ src/components/EditorBox/index.tsx
@@ -1,0 +1,4 @@
+import dynamic from "next/dynamic";
+export const EditorBox = dynamic(() => import("./EditorBox"), {
+  ssr: false,
+});
