+++ src/components/index.tsx
@@ -1,4 +1,1 @@
-import dynamic from "next/dynamic";
-export const EditorBox = dynamic(() => import("./EditorBox"), {
-  ssr: false,
-});
+export * from "./EditorBox";
